refactor: separate public API contract checker responsibilities
